% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_model_names.R
\name{assign_model_names}
\alias{assign_model_names}
\title{Give models a meaningful names for comparison.}
\usage{
assign_model_names(models)
}
\arguments{
\item{models}{}
}
\value{
a list contains the all the models with names.
}
\description{
Assign models with names based on method, metric, preProcess and sampling from the parameter of params dataframe in ml_lis funtion.
}
\details{
The function uses foreach function to get the method, metric,preProcess and sampling for each model from a list of models.
}
\examples{
down_sampling_models= assign_model_names(down_sampling_models)


}
\seealso{
ml_list, model_list_load
}
