% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoInterpret.R
\name{autoInterpret}
\alias{autoInterpret}
\title{Automated model interpretability}
\usage{
autoInterpret(train, trainedModel, sample = NULL, seed = 1991,
  verbose = TRUE)
}
\arguments{
\item{train}{[data.frame | Required] Training set on which the original model was trained}

\item{trainedModel}{[mlr model object | Required] A trained model using the mlr pacakge or produced via autoLearn}

\item{sample}{[numeric | Optional] A number between 0 - 1 to sub-sample the training set for faster computational time. Default of NULL which will result in a small sample}

\item{seed}{[integer | Optional] Random seed number for reproducable results. Default of 1991}
}
\value{
List containing plots
}
\description{
Generates various plots for model interpretability
}
\examples{
train <- iris
mod <- mlr::train(makeLearner("classif.ranger", predict.type = "prob"), iris.task)

plots <- autoInterpret(train = iris,
trainedModel = mod)
}
\author{
Xander Horn
}
