% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableLearners.R
\name{availableLearners}
\alias{availableLearners}
\title{Available models in autoML}
\usage{
availableLearners()
}
\value{
List of models
}
\description{
Returns a list of models to select from when using autoML
}
\examples{
availableLearners()
}
\author{
Xander Horn
}
