% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableMetrics.R
\name{availableMetrics}
\alias{availableMetrics}
\title{Available optimization metrics for machine learning models}
\usage{
availableMetrics()
}
\value{
List of metrics
}
\description{
Displays all available optimization metrics to be used within autoML
}
\examples{
 availableMetrics()
 @author
 Xander Horn
}
