% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_bs.R
\name{cal_bs}
\alias{cal_bs}
\title{Calculate Integrated Brier Score (IBS) for New Dataset}
\usage{
cal_bs(newdata, model, model_name)
}
\arguments{
\item{newdata}{Dataframe containing time-to-event data with 'time' and 'status' columns}

\item{model}{Trained survival analysis model object}

\item{model_name}{Model specification. Supported models:
"Lasso", "Ridge", "Enet", "RFSRC", "GBM", "CoxBoost", "plsRcox",
"XGBoost", "BlackBoost", "DeepHit", "DeepSurv", "SurvivalSVM","GLMBoost","SuperPC"}
}
\value{
Numeric Integrated Brier Score (IBS), where lower values indicate better performance
}
\description{
Computes Integrated Brier Score using trained survival analysis models
}
