% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_vali_index.R
\name{cal_vali_index}
\alias{cal_vali_index}
\title{Calculate AUC scores of Machine Learning Models in all data}
\usage{
cal_vali_index(
  list_train_vali_Data,
  candidate_genes,
  model_list,
  filter_OS_time = F,
  meta_time = "m",
  rep = 1,
  outdir = "4.test/"
)
}
\arguments{
\item{list_train_vali_Data}{A list contain the dataframe (colnames:ID,time,status,feature)}

\item{candidate_genes}{features that used to build the model}

\item{model_list}{A list contain trained model}

\item{filter_OS_time}{if keep survival time >30 days}

\item{meta_time}{follow up time used by days or months or years c("m","y","d") default m}

\item{rep}{repeats}

\item{outdir}{the output directory}
}
\value{
test.auc list containing:
        - model:
          - test data:
           - rep 1:
             - pred: risk score
             - cindex: c-index
             - bs : integrated brier score
             - auc : 1,2,3,5,7,10- year auc values
}
\description{
Calculate AUC scores of Machine Learning Models in all data
}
\examples{
\dontrun{
data(st_train_vali_Data)
test.auc<-cal_vali_index(list_train_vali_Data,
candidate_genes,
model_list,
rep=1,
outdir=out3)
}
}
