% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex_rank.R
\name{cindex_rank}
\alias{cindex_rank}
\title{C-index Distribution Visualization}
\usage{
cindex_rank(
  vali_auc_list,
  index,
  train = "TCGA",
  plot_type = "barplot",
  outdir = "./"
)
}
\arguments{
\item{vali_auc_list}{List of models containing performance metrics of all cohorts}

\item{index}{#cindex,bs,km_auc_1, km_auc_3, km_auc_5, km_auc_10}

\item{train}{# the training cohort default "TCGA"}

\item{plot_type}{Visualization type ("boxplot", "forestplot", or "barplot")}

\item{outdir}{Output directory for saving plots}
}
\value{
ggplot object or list of plots
}
\description{
Creates ranked performance plots for survival models
}
\examples{
\dontrun{
cindex_rank(
  vali_auc_list = test.index,
  index="cindex",
  train="TCGA",
  plot_type = "boxplot",
  outdir="./"
)
}
}
