% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeInteractions.R
\name{computeInteractions}
\alias{computeInteractions}
\title{Computes feature interactions}
\usage{
computeInteractions(x, numFeats, autoCode = TRUE, progress = FALSE)
}
\arguments{
\item{x}{[data.frame | Required] Data.frame containing numeric features to transform}

\item{numFeats}{[character vector | Required] Character vector of numerical features}

\item{autoCode}{[logical | Optional] Should code be generated when running the function}

\item{progress}{[logical | Optional] Should a progress bar display the progress when running the function}
}
\value{
List containing data.frame with interacted features as well as code when autoCode is TRUE
}
\description{
Computes feature interactions for numerical and integer features. Feature interactions include x1 + x2, x1 - x2, x1 * x2, x1 / x2
}
\examples{
int <- computeInteractions(x = iris, numFeats = names(iris)[1:4])
}
\author{
Xander Horn
}
