% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corefeature_selection.R
\name{Corefeature.Prog.Screen}
\alias{Corefeature.Prog.Screen}
\title{Screening out the core variables for the prognosis with the machine learning algorithms}
\usage{
Corefeature.Prog.Screen(
  InputMatrix,
  genelist,
  mode = NULL,
  seed = NULL,
  d = "m",
  single_ml = NULL,
  unicox_km = T
)
}
\arguments{
\item{InputMatrix}{A gene expression dataframe after log2(x+1) scaled. The first three of the column names are, in order, ID,OS.time, OS. Columns starting with the fourth are gene symbols. OS.time is a numeric variable in days. OS is a numeric variable containing 0, 1. 0: Alive, 1: Dead.}

\item{mode}{We provide three modes including 'all', 'single', and 'all_without_SVM'. The 'all' mode means using all eight methods for selecting. The 'single' mode means using only one method for running. Since SVM takes so much time, we're singling him out. The 'all_without_SVM' mode means the other seven methods used for selecting.}

\item{seed}{The seed. You can set it as any number. For example, 5201314.}

\item{single_ml}{The one method from the eight methods including "RSF", "Enet", "Boruta", "Xgboost", "SVM-REF", "Lasso", "CoxBoost', 'StepCox'.}

\item{candidate_genes}{The input genes, that you want to screen out from, for identifying the core features.}

\item{nodesize}{The node size parameter for the RSF method. The default is 5. You can try another positive integer. For example, 10,15,20, etc.}
}
\description{
A function can be used for screening out the core features from the given candidate genes with eight machine learning algorithms.
}
