% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folds.R
\name{create_folds}
\alias{create_folds}
\title{Create Repeated Stratified Folds for Cross-Validation}
\usage{
create_folds(data, fold, nrepeats, strata, seed)
}
\arguments{
\item{data}{A data frame containing the dataset to be split.}

\item{fold}{Integer, number of folds (k) to create.}

\item{nrepeats}{Integer, number of times to repeat the k-fold splitting.}

\item{strata}{Character, name of the column in `data` used for stratification.}

\item{seed}{Integer, random seed for reproducibility.}
}
\value{
A list of length `nrepeats`, where each element is a list of fold indices.
}
\description{
Generates repeated stratified cross-validation folds while preserving class distribution.
}
