% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFeatures.R
\name{distanceToCenter}
\alias{distanceToCenter}
\title{Computes unsupervised features on numerical and integer features}
\usage{
distanceToCenter(x, numFeats, clusters = 10, autoCode = TRUE,
  progress = FALSE, seed = 1234)
}
\arguments{
\item{x}{[data.frame | Required] Data.frame containing numeric features to transform}

\item{numFeats}{[character vector | Required] Character vector of numerical features}

\item{clusters}{[integer, Optional] Number of clusters to create}

\item{autoCode}{[logical | Optional] Should code be generated when running the function}

\item{progress}{[logical | Optional] Should a progress bar display the progress when running the function}

\item{seed}{[integer | Optional] Random seed number for reproducable results. Default of 1991}
}
\value{
List containing data.frame with clustered features as well as code when autoCode is TRUE
}
\description{
K-means clustering is used to create new features on existing numeric and integer features, then calculating the distance to center and using this as the new feature
}
\examples{
clst <- distanceToCenter(x = iris, numFeats = names(iris)[1:4])
}
\author{
Xander Horn
}
