% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_one_hot.R
\name{encode_one_hot}
\alias{encode_one_hot}
\title{One-hot encoding of chategorical or caracter variables.}
\usage{
encode_one_hot(
  data,
  remove_selected_columns = TRUE,
  select_columns = NULL,
  remove_first_dummy = FALSE,
  remove_most_frequent_dummy = FALSE,
  ignore_na = FALSE,
  split = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{remove_selected_columns}{If TRUE (not default), removes the columns
used to generate the dummy columns.}

\item{select_columns}{Vector of column names that you want to create dummy
variables from. If NULL (default), uses all character and factor columns.}

\item{remove_first_dummy}{Removes the first dummy of every variable such
that only n-1 dummies remain. This avoids multicollinearity issues in models.}

\item{remove_most_frequent_dummy}{Removes the most frequently observed
category such that only n-1 dummies remain. If there is a tie for most
frequent, will remove the first (by alphabetical order) category that is
tied for most frequent.}

\item{ignore_na}{If TRUE, ignores any NA values in the column.
If FALSE (default), then it will make a dummy column for value_NA
and give a 1 in any row which has a NA value.}

\item{split}{A string to split a column when multiple categories are in the
cell. For example, if a variable is Pets and the rows are "cat",
"dog", and "turtle", each of these pets would become its own dummy column.
If one row is "cat, dog", then a split value of "," this row would have a
value of 1 for both the cat and dog dummy columns.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
}
\examples{
\dontrun{

my_data <- encode_one_hot(my_data)
}
}
\author{
Resul Akay
}
