% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeMean.R
\name{encodeMean}
\alias{encodeMean}
\title{Mean encoding for categorical features}
\usage{
encodeMean(x, y, catFeats, autoCode = TRUE)
}
\arguments{
\item{x}{[data.frame | Required] Data.frame containing categorical features}

\item{y}{[character | Required] Target feature name}

\item{catFeats}{[character vector | Required] Character vector of categorical features}

\item{autoCode}{[logical | Optional] Should code be generated when running the function}
}
\value{
List containing data.frame with encoded features as well as code when autoCode is TRUE
}
\description{
Uses the target feature to calculate the mean for each level of a categorical feature and encodes it
}
\author{
Xander Horn
}
