% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeMissing.R
\name{encodeMissing}
\alias{encodeMissing}
\title{Encode missing values in features}
\usage{
encodeMissing(x, numEncode = NULL, charEncode = c("", " ", "UNKNOWN",
  "MISS", "MISSING", "UNK", "NA", "NULL", "N/A"), autoCode = TRUE)
}
\arguments{
\item{x}{[data.frame | Required] Data.frame containing numeric features to transform}

\item{numEncode}{[numeric vector | Optional] Numeric vector containing reserved values for missing data. Default of NULL}

\item{charEncode}{[character vector | Optional] Character vector containing reserved values for missing data. Default of c('',' ','UNKNOWN','MISS','MISSING','UNK','NA','NULL','N/A')}

\item{autoCode}{[logical | Optional] Should code be generated when running the function}
}
\value{
List containing data.frame with encoded features as well as code when autoCode is TRUE
}
\description{
Changes the way in which missing values are represented within features. Various systems uses different reserved values for missing data, this function uses the most common reserved values and replaces them with NA
}
\author{
Xander Horn
}
