% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeOneHot.R
\name{encodeOneHot}
\alias{encodeOneHot}
\title{One hot encoding of categorical features}
\usage{
encodeOneHot(x, catFeats, autoCode = TRUE)
}
\arguments{
\item{x}{[data.frame | Required] Data.frame containing categorical features}

\item{catFeats}{[character vector | Required] Character vector of categorical features}

\item{autoCode}{[logical | Optional] Should code be generated when running the function}
}
\value{
List containing data.frame with encoded features as well as code when autoCode is TRUE
}
\description{
Performs one hot encoding for categorical features
}
\examples{
temp <- encodeOneHot(x = iris, catFeats = "Species")
}
\author{
Xander Horn
}
