% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeOrdinal.R
\name{encodeOrdinal}
\alias{encodeOrdinal}
\title{Ordinal encoding of categorical features}
\usage{
encodeOrdinal(x, catFeats, autoCode = TRUE)
}
\arguments{
\item{x}{[data.frame | Required] Data.frame containing categorical features}

\item{catFeats}{[character vector | Required] Character vector of categorical features}

\item{autoCode}{[logical | Optional] Should code be generated when running the function}
}
\value{
List containing data.frame with encoded features as well as code when autoCode is TRUE
}
\description{
Encodes categorical features using a ordinal approach where the proportion is first calculated, sorted and then a transformation is applied to the sorted proportions
}
\examples{
temp <- encodeOrdinal(x = iris, catFeats = "Species")
}
\author{
Xander Horn
}
