% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeProportion.R
\name{encodeProportion}
\alias{encodeProportion}
\title{Proportional encoding of categorical features}
\usage{
encodeProportion(x, catFeats, autoCode = TRUE)
}
\arguments{
\item{x}{[data.frame | Required] Data.frame containing categorical features}

\item{catFeats}{[character vector | Required] Character vector of categorical features}

\item{autoCode}{[logical | Optional] Should code be generated when running the function}
}
\value{
List containing data.frame with encoded features as well as code when autoCode is TRUE
}
\description{
Encodes categorical features using the proportional approach
}
\examples{
temp <- encodeProportion(x = iris, catFeats = "Species")
}
\author{
Xander Horn
}
