% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engineerDateTime.R
\name{engineerDateTime}
\alias{engineerDateTime}
\title{Engineer date features}
\usage{
engineerDateTime(x, datetimeFeats, autoCode = TRUE)
}
\arguments{
\item{x}{[data.frame | Required] Data.frame containing numeric features to transform}

\item{datetimeFeats}{[character vector | Required] Character vector of date/ datetime features. Date time features cannot have the same value}

\item{autoCode}{[logical | Optional] Should code be generated when running the function}
}
\value{
List containing data.frame with date engineered features as well as code when autoCode is TRUE
}
\description{
Creates new featues based off of existing date features and then drops the date features after engineering has taken place. Features created include year, month, week, day, weekday, hour, minute, second
}
\examples{
x <- iris
x$date <- Sys.Date() + as.integer(row.names(x))
feats <- engineerDateTime(x = x, datetimeFeats = "date")
}
\author{
Xander Horn
}
