% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engineerTextFeatures.R
\name{engineerText}
\alias{engineerText}
\title{Automated feature engineering of text features}
\usage{
engineerText(x, textFeats, autoCode = TRUE)
}
\arguments{
\item{x}{[data.frame | Required] Data.frame containing text features}

\item{textFeats}{[character vector | Required] Character vector containing text feature names}

\item{autoCode}{[logical | Optional] Should code be generated when running this function}
}
\description{
Automatically creates new features for all text features provided. Used for NLP cases
}
\author{
Xander Horn
}
