% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eusilc2011.R
\docType{data}
\name{eusilc2011}
\alias{eusilc2011}
\title{Austrian Income and Living Conditions data (EU-SILC Austria (2011))}
\format{
An object of class \code{data.frame} with 5880 rows and 19 columns.
}
\source{
\url{https://ec.europa.eu/eurostat/cache/microdata/eusilc/AT_2011_EUSILC.zip}
}
\usage{
eusilc2011
}
\description{
Austrian Income and Living Conditions data (EU-SILC Austria (2011))
}
\examples{
\dontrun{

fit <- lm(rent_ful ~ region+dwel_type+ tenure_status+leaking_roof+
bath_room+indoor_wc+dark_home+noisy_home+ bad_surrounding+
crime_surrounding+rooms,data = eusilc2011)
}
}
\keyword{datasets}
