% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metrics.R
\name{extract_metrics}
\alias{extract_metrics}
\title{Calculate Average Performance Metrics for Cross-Validated Models}
\usage{
extract_metrics(model_list, metric_names = NULL)
}
\arguments{
\item{model_list}{List of models containing train/valid/test performance metrics}

\item{metric_names}{Vector of metric names to extract (default includes C-index,
IBS and time-dependent AUCs)}
}
\value{
List containing three data frames:
        - train: Training set metrics
        - valid: Validation set metrics
        - test: Test set metrics
}
\description{
Computes mean performance metrics across training, validation and test sets
from multiple cross-validated models
}
