% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastEDA.R
\name{fastEDA}
\alias{fastEDA}
\title{Fast exploratory data analysis}
\usage{
fastEDA(x, numChars = 25, seed = 1991)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
Fast exploratory data analysis
}
\examples{

}
