% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{feature_selection}
\alias{feature_selection}
\title{Screening out the core prognostic genes with the machine learning algorithms}
\usage{
feature_selection(
  InputMatrix,
  genelist = NULL,
  fold = 5,
  filter_OS_time = F,
  meta_time = "m",
  unicox_km = T,
  deg = T,
  up = F,
  method = "all",
  svm_method = F,
  outdir = "1.feature_select/",
  seed = 123,
  ncore = 4
)
}
\arguments{
\item{InputMatrix}{### 1: ID, 2: OS_time, 3: OS_status (0/1), feature list...}

\item{genelist}{The candidate genes}

\item{fold}{default 5}

\item{filter_OS_time}{default F If keep follow up time > 30 days}

\item{meta_time}{When filter_OS_time=T, set follow up time months(m) or days(d) or years(y)}

\item{unicox_km}{unicox and km filtering}

\item{deg}{differential expression filtering}

\item{up}{when up=T, HR>1 in unicox analysis and log2FC>0 in differential analysis}

\item{method}{"all" using 9 ML algorithms or boruta or xgboost or RSF or rfe or lasso}

\item{svm_method}{when method="all", setting if using svm algorithms, which is time-consuming}

\item{outdir}{the outpt directory}

\item{seed}{The seed}

\item{ncore}{multisession workers}
}
\value{
feature list
}
\description{
A function can be used for screening out the core features from the given candidate genes with 9 machine learning algorithms.
}
\examples{
\donttest{
# Requires trained data and genelist
data(train_data)
selected.feature<-feature_selection(InputMatrix,genelist=genelist)
}
}
