% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection2.R
\name{feature_selection2}
\alias{feature_selection2}
\title{Screening out the core prognostic genes with the machine learning algorithms}
\usage{
feature_selection2(
  InputMatrix,
  genelist,
  filter_OS_time = F,
  meta_time = "d",
  deg = T,
  up = F,
  outdir = "1.feature_select/",
  seed = 123,
  height = 8,
  width = 10
)
}
\arguments{
\item{InputMatrix}{### 1: ID, 2: OS_time, 3: OS_status (0/1), feature list...}

\item{genelist}{The candidate genes}

\item{filter_OS_time}{default F If keep follow up time > 30 days}

\item{meta_time}{When filter_OS_time=T, set follow up time months(m) or days(d) or years(y)}

\item{deg}{differential expression filtering}

\item{up}{when up=T, HR>1 in unicox analysis and log2FC>0 in differential analysis}

\item{outdir}{the output directory}

\item{seed}{The seed}

\item{height}{8}

\item{width}{10}
}
\value{
feature list
}
\description{
A function can be used for screening out the core features from the given candidate genes with 9 machine learning algorithms.
}
\examples{
\donttest{
# Requires trained data and genelist
data(train_data)
selected.feature<-feature_selection2(InputMatrix,genelist=genelist)
}
}
