% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureImportance.R
\name{featureImportance}
\alias{featureImportance}
\title{Measures feature importance for any model provided}
\usage{
featureImportance(train, trainedModel, seed = 1991, topFeats = NULL,
  sample = 0.1, outputPath = NULL)
}
\arguments{
\item{train}{[data.frame | Required] Training set on which the model was trained}

\item{trainedModel}{[mlr obj | Required] MLR trained moodel object}

\item{seed}{[integer | Optional] Random seed number for reproducable results. Default of 1991}

\item{sample}{[numeric | Optional] A number between 0 - 1 to sub-sample the training set for faster computational time. Default of 0.1}
}
\value{
List object containing a data.frame and a plot object.
}
\description{
Measures feature importance for any model provided
}
\examples{
mod <- mlr::train(makeLearner("classif.ranger"), iris.task)
featureImportance(train = iris, mod)
}
\author{
Xander Horn
}
