% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixCatProportions.R
\name{fixCatProportions}
\alias{fixCatProportions}
\title{Correct low proportional levels of categorical features}
\usage{
fixCatProportions(x, catFeats, minLevelPercentage = 0.025,
  autoCode = TRUE)
}
\arguments{
\item{x}{[data.frame | Required] Data.frame containing categorical features}

\item{catFeats}{[character vector | Required] Character vector of categorical features}

\item{minLevelPercentage}{[numeric | Optional] Used to identify low proportional categorical levels. Default of 0.025}

\item{autoCode}{[logical | Optional] Should code be generated when running the function}
}
\value{
List containing a data.frame with and a code list if autoCode is set to TRUE
}
\description{
Corrects low proportions levels by investigating the proportion of data allocated to each level and replacing the level by a one called "_ALL_OTHER_" if the level is below the minimum percentage threshold
}
\examples{
new <- fixCatProportions(x = iris, catFeats = "Species") # Proportions are all at 0.33 thus no change will occur
}
\author{
Xander Horn
}
