% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_boxplot.R
\name{generate_two_boxplot}
\alias{generate_two_boxplot}
\title{boxplot of differential expression between two groups}
\usage{
generate_two_boxplot(
  expr_matrix,
  genelist,
  grouping,
  orders = NULL,
  dataset_name = NULL,
  method = "t.test",
  width = 5.5,
  height = 7,
  outdir = "gene_expression/boxplot/"
)
}
\arguments{
\item{expr_matrix}{expr dataframe col features rows accession}

\item{genelist}{a gene list}

\item{grouping}{grouping datafarme}

\item{orders}{order for group}

\item{dataset_name}{title}

\item{method}{"t.test" or "wilcox"}

\item{width}{plot width}

\item{height}{plot height}

\item{outdir}{the output directory}
}
\description{
boxplot of differential expression between two groups
}
