% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBalancedTask.R
\name{generateBalancedTask}
\alias{generateBalancedTask}
\title{Generates a balanced task using oversampling from an existing mlr task}
\usage{
generateBalancedTask(task)
}
\arguments{
\item{task}{}
}
\description{
Generates a balanced task using oversampling from an existing mlr task
}
