% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHyperParams.R
\name{generateHyperParams}
\alias{generateHyperParams}
\title{Generates a list of hyper parameters suitable for specific learners}
\usage{
generateHyperParams(learners = NULL, task, clusters = NULL,
  cores = NULL)
}
\description{
Generates a list of hyper parameters suitable for specific learners
}
