% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLearners.R
\name{generateLearners}
\alias{generateLearners}
\title{Generates a list of mlr learners}
\usage{
generateLearners(task)
}
\arguments{
\item{task}{}
}
\value{

}
\description{
Generates a list of mlr learners
}
