% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMetrics.R
\name{generateMetrics}
\alias{generateMetrics}
\title{Generates optimization metrics}
\usage{
generateMetrics(task)
}
\arguments{
\item{task}{}
}
\value{

}
\description{
Creates a list of suitable metrics that can be used to train a model
}
