% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggforest2.R
\name{ggforest2}
\alias{ggforest2}
\title{Forest Plot for Cox Proportional Hazards Model}
\usage{
ggforest2(
  model,
  data = NULL,
  main = "Hazard ratio",
  cpositions = c(0.02, 0.1, 0.35),
  fontsize = 0.7,
  refLabel = "reference",
  noDigits = 2
)
}
\arguments{
\item{model}{an object of class coxph.}

\item{data}{a dataset used to fit survival curves. If not supplied then data
will be extracted from 'fit' object.}

\item{main}{title of the plot.}

\item{cpositions}{relative positions of first three columns in the OX scale.}

\item{fontsize}{relative size of annotations in the plot. Default value: 0.7.}

\item{refLabel}{label for reference levels of factor variables.}

\item{noDigits}{number of digits for estimates and p-values in the plot.}
}
\value{
returns a ggplot2 object (invisibly)
}
\description{
Drawing Forest Plot for Cox proportional hazards model. In two panels the model structure is presented.
}
