% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex_sig_comp.R
\name{indexC_comp}
\alias{indexC_comp}
\title{Compare the AUC of our model with previously published models}
\usage{
indexC_comp(own_auc_list, published_auc_list = NULL, model_name, dataset)
}
\arguments{
\item{own_auc_list}{Output of function cal_vali_index}

\item{published_auc_list}{Output of function cal_vali_index}

\item{model_name}{Model name used to compare}

\item{dataset}{A vector of names for all datasets}
}
\value{
a list
}
\description{
Creates a distribution plot of AUC among different mdoels
}
