% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pkg_model_list.R
\name{install_pkg_model_list}
\alias{install_pkg_model_list}
\title{install missing packages from pre-trained models}
\usage{
install_pkg_model_list(models)
}
\arguments{
\item{models}{A list of pre-trained caret models.}
}
\description{
check and install missing packages dependencies from a list of pre-trained caret models.
}
\details{
Given a list of pre-trained caret models from others, you might not have all the packages required to predict with those models. That is when you should you use this function.
The input of the function is just a list of caret models. This function depends on the function install_pkg_model_names
}
\examples{
install_pkg_model_list(churn_down_sampling_models)

install_pkg_model_list(models=down_sampling_models)


}
