% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localModelInterpretability.R
\name{localModelInterpretability}
\alias{localModelInterpretability}
\title{Local model interpretation}
\usage{
localModelInterpretability(train, trainedModel, seed = 1991,
  sample = 0.1)
}
\arguments{
\item{train}{[data.frame | Required] Training set on which the model was trained}

\item{trainedModel}{[mlr obj | Required] MLR trained moodel object}

\item{seed}{[integer | Optional] Random seed number for reproducable results. Default of 1991}

\item{sample}{[numeric | Optional] A number between 0 - 1 to sub-sample the training set for faster computational time. Default of 0.1}
}
\value{
List object containing a data.frame and a plot object.
}
\description{
Fits a linear regression model to the predicted responses or probabilities of another model. Note that for classification models, the model must predict a probability else this fails.
}
\examples{
mod <- mlr::train(makeLearner("classif.ranger", predict.type = "prob"), iris.task)
localModelInterpretability(train = iris, mod)
}
\author{
Xander Horn
}
