% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.survival.model.R
\name{ML.survival.model}
\alias{ML.survival.model}
\title{Training survival model by using 13 Machine Learning algorithms}
\usage{
ML.survival.model(
  train_data,
  candidate_genes,
  filter_OS_time = F,
  meta_time = "m",
  cor = F,
  cor_threshold = 0.85,
  fold = 5,
  rep = 10,
  p = 1,
  deep = F,
  outdir = "2.train/",
  seed = 123,
  ncore = 4
)
}
\arguments{
\item{train_data}{# used for training (colnames:ID,time,status,feature)}

\item{candidate_genes}{features that used to build the model}

\item{filter_OS_time}{if keep survival time >30 days}

\item{meta_time}{follow up time used by days or months or years c("m","y","d") default m}

\item{cor}{if remove highly correlated genes}

\item{cor_threshold}{correlation 0.85}

\item{fold}{folds for cross validation}

\item{rep}{repeats}

\item{p}{A proportion of train_data used to build the model default 0.7}

\item{deep}{if using deep learning model deepsurv or deephit}

\item{outdir}{the output directory}

\item{seed}{random seed}

\item{ncore}{multisession cores}
}
\value{
model list
}
\description{
Training survival model by using 13 Machine Learning algorithms
}
\examples{
model_list <- ML.survival.model(
  train_data,
  candiate_genes,
  filter_OS_time=F,
  meta_time="m",
  mode="all",
  seed=123,
  fold=5,
  rep=10,
  p=0.7
)

}
