% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_bwplot.R
\name{ml_bwplot}
\alias{ml_bwplot}
\title{A function to visualize the model performance.}
\usage{
ml_bwplot(models, metric = NULL)
}
\arguments{
\item{models}{A list of models from ml_list function or manually combine finely tuned models from ml_tune or train function from caret package.}

\item{metric}{A character, the metric of model performance, the available values are "ROC","Sens","Spec","Accuracy","Kappa".}
}
\description{
A wrapper around the bwplot function in caret to visualize models even if models are trained using different metrics.
}
\details{
If the models contains metrics from c("ROC","Sens","Spec") and c("Accuracy","Kappa"), then it will plot two graphs.
 You could not plot the "ROC","Sens" or "Spec" value for a model if the model trained by using "Accuracy" or "Kappa", and vice versa.
}
\examples{
# plot all the metrics
\dontrun{
  ml_bwplot(models)
}

# plot only the ROC metric for all models.
\dontrun{
ml_bwplot(models,metric="ROC")
}


}
\seealso{
To see how the list of models are generated \code{\link{ml_list}} and \code{\link{ml_tune}}.
 The bwplot function is not documented on CRAN but you could see it on caret tutorial \url{https://topepo.github.io/caret/model-training-and-tuning.html}
}
