% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_cor_filter.R
\name{ml_cor_filter}
\alias{ml_cor_filter}
\title{A function to filter models based on inter-model correlation. Mainly based on modelCor function in caret.}
\usage{
ml_cor_filter(models, cor_level = 0.9)
}
\arguments{
\item{models}{A list of caret models generated from ml_list function or by combining caret models into a list manually.}

\item{cor_level}{A numeric, the maximum inter-model correlation computed from modelCor function from caret package.}
}
\value{
A list of caret models that its performance metrics from cross-validation satisfy the conditions.
}
\description{
This function will also remove models that have NA value in resampled performance. a.k.a NA value in modelCor function output.
}
\details{
What the function does is to first get rid of model performance with missing values. Second, filter models based on cor_level,
which is to get rid of the models with high correlation.
}
\examples{
\dontrun{
 # a maximum of correlation 0.9 between models.
 low_cor_models_churn=testmodels_churn\%>\%ml_cor_filter(cor_level = 0.9)
 # a maximum of correlation 0.7 between models.
 low_cor_models_churn=testmodels_churn\%>\%ml_cor_filter(cor_level = 0.7)

}

 @export


}
