% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_list_load.R
\name{model_list_load}
\alias{model_list_load}
\title{A function to load all rds files in a specified folder from disk.}
\usage{
model_list_load(path)
}
\arguments{
\item{path}{The path of the folder}
}
\value{
A list of models saved to disk. Same output as ml_list function.
}
\description{
This function is used to load all models saved to disk in ml_list function but could be used to load all rds files in a folder.
}
\details{
Each rds file will be added to a list that is returned by this function.
}
\examples{

\dontrun{
   models=model_list_load(path="~/Dropbox/churn/down_sampling")
}
}
