% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyCode.R
\name{modifyCode}
\alias{modifyCode}
\title{Modify production code to be model specific}
\usage{
modifyCode(trainedModel, edaFrame, codeFrame)
}
\arguments{
\item{trainedModel}{[mlrObject | Required] A trained model from the mlr R library or models returned from autoML / autoLearn}

\item{edaFrame}{[data.frame | Required] EDA frame returned by autoPreProcess}

\item{codeFrame}{[data.frame | Required] Code frame returned by autoPreProcess}
}
\value{
Data.frame object containing the modified code
}
\description{
Modifies the production code generated by the autoPreProcess function to be specific to a trained model object. Results in less code and only uses the features used in the final trained model. The saveCode function can then be used to save the code.
}
\author{
Xander Horn
}
