% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.auto_xgb.R
\name{predict.auto_xgb}
\alias{predict.auto_xgb}
\title{Predict an auto_xgboost object.}
\usage{
\method{predict}{auto_xgb}(object, newdata, ...)
}
\arguments{
\item{object}{an auto_xgboost object.}

\item{newdata}{newdata}

\item{\dots}{Other arguments}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
}
\author{
Resul Akay
}
