% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_data_combine.R
\name{run_data_combine}
\alias{run_data_combine}
\title{Data preprocessing for model performance evaluation}
\usage{
run_data_combine(
  list_data_surv,
  sets = NULL,
  correct_method = "none",
  outdir = "dataset/"
)
}
\arguments{
\item{list_data_surv}{#column 1 ID # 2 survival time # 3 status 1 or 0 # 4 gene variables}

\item{sets}{c(1,2,3,4)}

\item{correct_method}{sva,scale,min_max,none}

\item{outdir}{output path}
}
\value{
data list
}
\description{
Data preprocessing for model performance evaluation
}
