% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_meta_combine.R
\name{run_meta_combine}
\alias{run_meta_combine}
\title{Data preprocessing for model performance evaluation}
\usage{
run_meta_combine(
  list_data_surv,
  sets = NULL,
  features = c("Sample_ID", "Age", "PSA", "GS", "path_T"),
  outdir = "dataset/"
)
}
\arguments{
\item{list_data_surv}{#column 1 Sample_ID # 2 clinical features}

\item{sets}{c(1,2,3,4)}

\item{features}{clinical variables}

\item{outdir}{output path for example "Age","PSA","GS","path_T"}
}
\value{
data list
}
\description{
Data preprocessing for model performance evaluation
}
