% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveCode.R
\name{saveCode}
\alias{saveCode}
\title{Save automatically generated code}
\usage{
saveCode(codeFrame, filename = "autoCode", path = NULL)
}
\arguments{
\item{codeFrame}{[data.frame | Required] A data.frame object produced by autoPreProcess}

\item{filename}{[character | Optional] The name of the file in which the code will be saved}

\item{path}{[character | Optional] The path where the file should be saved containing the code}
}
\value{
Local file with code
}
\description{
Saves code generated from autoPreProcess to local file
}
\author{
Xander Horn
}
