% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaler.R
\name{scaler}
\alias{scaler}
\title{Scales a numeric feature using the max approach}
\usage{
scaler(df, numFeats, autoCode = TRUE)
}
\arguments{
\item{df}{[data.frame | Required] Data.frame object containing numeric features}

\item{numFeats}{[character vector | Required] Character vector specifying the numeric features on which scaling should be applied}

\item{autoCode}{[logical | Optional] Should code be produced when running this function}
}
\value{
List object containng a data.frame with scaled numeric features and all other features contained in the dataset as well as a list containing code
}
\description{
Scales a numeric feature using the max approach
}
\examples{
new_iris <- scaler(df = iris, numFeats = names(iris)[1:4])
}
\author{
Xander Horn
}
