% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchFeatureFormatting.R
\name{searchFeatFormatting}
\alias{searchFeatFormatting}
\title{Feature formatting searching}
\usage{
searchFeatFormatting(x, minUnique = 25, minChrPercentage = 0.2,
  seed = 1991)
}
\arguments{
\item{x}{[data.frame | Required] Data.frame containing features}

\item{minUnique}{[integer | Optional] Integer specifying the minimum unique values of a feature in order to format the feature as a categorical or numeric. Deafult of 25}

\item{minChrPercentage}{[numeric | Optional] Searches a character feature for a minimum percentage of character values in order to identify incorrectly formatted numeric features. Default of 0.2}

\item{seed}{[integer | Optional] Random seed for reproducable results. Default of 1991}
}
\value{
Returns a data.frame containing the feature name, original class and recommended class
}
\description{
Seraches for the best formats for features in the provided dataset
}
