% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeRecordB.R
\name{timeRecordB}
\alias{timeRecordB}
\title{A function to record the time used in long time calculation.}
\usage{
timeRecordB(output_message = "None")
}
\arguments{
\item{output_message}{A string that contains the message you want to saved.}
}
\description{
This function record the current time of when the function is used, and write that information into a txt file.
}
\details{
An global variable is generated in the session,
if this variable does not exist in the current R session. Then the output_message and the current time is written to a
file with the name that contains the current time. This function does not allow to to specify
the current time on purpose, but if you want to continue to append to one log file, you could assign the variable
with the file name in the global environment.
}
\examples{
timeRecordB()
sum(1:10)
timeRecordB(output_message="sum from 1 to 10")

extremely_long_name_i_do_not_think_anyone_would_be_sanely_to_use="logfile_names_to_append.log"
timeRecordB()
sum(1:10)
timeRecordB(output_message="sum from 1 to 10")



}
