% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeRecordR.R
\name{timeRecordR}
\alias{timeRecordR}
\title{A function to read the time record in the timeRecordB function.}
\usage{
timeRecordR(unit = "s", ignore = 1)
}
\arguments{
\item{unit}{A character. The unit for the time.}

\item{ignore}{A numeric. The minimal time interval to include.}
}
\value{
dataframe for the recorded time intervals with output message .
}
\description{
This function read the time record from the log file created by timeRecordB.
}
\details{
Given the current global variable extremely_long_name_i_do_not_think_anyone_would_be_sanely_to_use,
timeRecordR function would try to read in the log file, then time interval is calculated for each recorded time.
}
\examples{

timeRecordR()


library(dplyr)
timeRecordR()\%>\%filter(output_message!="None")\%>\%select(output_message,run_time)

}
