% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeExplainer.R
\name{treeExplainer}
\alias{treeExplainer}
\title{Decision tree local model explainer}
\usage{
treeExplainer(train, trainedModel, sample = 0.1, seed = 1991,
  maxDepth = 2)
}
\arguments{
\item{train}{[data.frame | Required] Training set on which the original model was trained}

\item{trainedModel}{[mlr model object | Required] A trained model using the mlr pacakge or produced via autoLearn}

\item{sample}{[numeric | Optional] A number between 0 - 1 to sub-sample the training set for faster computational time. Default of 0.1}

\item{seed}{[integer | Optional] Random seed number for reproducable results. Default of 1991}

\item{maxDepth}{[integer | Optional] Max depth of the decision tree. Default of 2}
}
\value{
Returns a list containing a plot
}
\description{
Fits a decision tree model to another model's predictions with the aim of explaining a black box model with decision rules
}
\examples{
mod <- mlr::train(makeLearner("classif.ranger"), iris.task)
treeExplainer(train = iris, mod, maxDepth = 5)
}
\author{
Xander Horn
}
