% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unload_extra_library.R
\name{unload_extra_library}
\alias{unload_extra_library}
\title{Unload additional packages from search path.}
\usage{
unload_extra_library(init_state = init_state)
}
\arguments{
\item{init_state}{A list of character vectors containing the information of previous state of search path.}
}
\description{
Some packages would load with itself some other package depedency into the search path. In order to eliminate the possibility of conflict
and avoid side effect in writing functions, it is recommended to remove additional packages from the search path after the calculation.
}
\details{
This is also a dirty trick to remove extra packages from search path when you are writing a package and
some functions from other packages you use have implicit or ill-documented depedencies so that you have to use require or library somewhere in your code.
}
\examples{
init_state=search()
testing_packs=function(data){
  library(caret)
  library(doParallel)
  library(foreach)
  library(magrittr)
  library(smotefamily)
  library(FNN)
  library(dbscan)
  library(ROSE)
  library(DMwR)
  library(h2o)
  library(dplyr)
  library(glue)
  library(MLmetrics)
  library(testthat)
  return(1)
}
testing_packs(data=iris)
unload_extra_library(init_state=init_state)
}
